The physics work!
Not only did we make it so you can jump and move reliably, it feels pretty good to play. The speeds need slight adjusting probably, but that can be done.

Next goal is probably hooking up the jump animation to the player when they jump. That might be hard considering the fact that I put the whole jump animation into one thing, instead of separating it into jump and fall. Also hook up the walk animation.