Simple Collision Fixes & Progress on Other Animations
A few things happened here. Firstly, I used a tutorial to make an Animator Controller to try and get a walking animation working. In the process, I renamed a few files for ease of access and added isMoving variable to the PlayerController (it doesn't have functionality yet). Also, I shrunk the player a little because of the fact that they could easily get stuck between tiles. This resulted in me modifying the jump strength greatly, so that you could actually see the ground as you fell back down.

So in summary:
- Added animator controller along with isMoving variable

Fixes:
- Shrank player so they can't get stuck between 1 block gaps
- Modified jump height to be more realistic and fun + so you could see the ground when jumping.

Do next:
- Fix jumping & movement. Can still jump on walls and quickly go up them, sometimes you jump really high for some reason, and momentum is halted when jump begins. You also still bounce a little when moving.
- Finish implementing isMoving variable and add more animations.